# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ExportRequest": {
        "additionalProperties": false,
        "properties": {
          "format": {
            "const": "csv",
            "default": "csv",
            "title": "Format",
            "type": "string"
          }
        },
        "title": "ExportRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortfolioItem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "latest_published_report_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Report Date"
          },
          "latest_published_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Snapshot Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reporting_currency": {
            "title": "Reporting Currency",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "reporting_currency",
          "latest_published_report_date",
          "latest_published_snapshot_id"
        ],
        "title": "PortfolioItem",
        "type": "object"
      },
      "PortfolioListResponse": {
        "additionalProperties": false,
        "properties": {
          "combined_report_dates": {
            "items": {
              "type": "string"
            },
            "title": "Combined Report Dates",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PortfolioItem"
            },
            "title": "Items",
            "type": "array"
          },
          "report_date_mismatch": {
            "title": "Report Date Mismatch",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "combined_report_dates",
          "report_date_mismatch"
        ],
        "title": "PortfolioListResponse",
        "type": "object"
      },
...
      "ReportArtifact": {
        "additionalProperties": false,
        "properties": {
          "artifact_sha256": {
            "title": "Artifact Sha256",
            "type": "string"
          },
          "artifact_url": {
            "title": "Artifact Url",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "disclosures": {
            "items": {
              "type": "string"
            },
            "title": "Disclosures",
            "type": "array"
          },
          "format": {
            "const": "csv",
            "title": "Format",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "snapshot_id",
          "format",
          "requested_by",
          "artifact_sha256",
          "disclosures",
          "created_at",
          "artifact_url"
        ],
        "title": "ReportArtifact",
        "type": "object"
      },
      "ReportListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportArtifact"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ReportListResponse",
        "type": "object"
      },
...
      "SnapshotListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SnapshotListResponse",
        "type": "object"
      },
...
      "SnapshotSummary": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "portfolio": {
            "title": "Portfolio",
            "type": "string"
          },
          "report_date": {
            "title": "Report Date",
            "type": "string"
          },
          "source_upload_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Upload Id"
          },
          "status": {
            "enum": [
              "draft",
              "validating",
              "validated",
              "approved",
              "published",
              "failed",
              "rejected",
              "superseded",
              "withdrawn"
            ],
            "title": "Status",
            "type": "string"
          },
          "value_label": {
            "title": "Value Label",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "import_id",
          "portfolio",
          "report_date",
          "version",
          "status",
          "value_label"
        ],
        "title": "SnapshotSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OidcBearer": {
        "description": "Production OIDC access token. Development headers are local-only.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Portfolio Operations Insight API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/portfolios": {
      "get": {
        "operationId": "listPortfolios",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "List governed portfolios and publication dates",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/portfolios/{code}/snapshots": {
      "get": {
        "operationId": "listPortfolioSnapshots",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_unpublished",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unpublished",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_superseded",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Superseded",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "List portfolio snapshots",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/reports/{report_id}/artifact": {
      "get": {
        "operationId": "downloadReportArtifact",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Download an exact report artifact",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/v1/snapshots/{snapshot_id}/reports": {
      "get": {
        "operationId": "listSnapshotReports",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "List immutable report runs",
        "tags": [
          "reports"
        ]
      },
      "post": {
        "operationId": "createSnapshotReport",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportArtifact"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Generate a controlled operational export",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}